% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cie1995_13.3}
\alias{cie1995_13.3}
\title{CIE-1974 test-colour samples}
\format{
A wide format data frame with 15 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(360, 830, 5)}}
\item{TCS_01}{CIE test colour sample 01, relativ spectra}
...
\item{TCS_14}{CIE test colour sample 14, relativ spectra}
}
}
\source{
\insertRef{CIE1995}{lighting}
}
\usage{
cie1995_13.3
}
\description{
\loadmathjax
A dataset of spectral radiance factor of the CIE-1974 test-colour samples to
be used in calculating the General Colour Rendering Index.
}
\keyword{datasets}
