% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cie2018_015}
\alias{cie2018_015}
\title{CIE daylight components}
\format{
A wide format data frame with 4 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(300, 830, 5)}}
\item{S_0}{CIE daylight component 0, relativ spectra}
\item{S_1}{CIE daylight component 1, relativ spectra}
\item{S_2}{CIE daylight component 2, relativ spectra}
}
}
\source{
\insertRef{CIE1995}{lighting}
}
\usage{
cie2018_015
}
\description{
A dataset containing the components of the relative spectral distribution of
daylight used in the calculation of relative spectral power distributions of
CIE daylight illuminants of different correlated colour temperatures.
}
\keyword{datasets}
