% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cie2018_S_026}
\alias{cie2018_S_026}
\title{CIE alpha-opic action spectra}
\format{
A wide format data frame with 4 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(300, 830, 5)}}
\item{s_sc}{S-cone-opic relating to the human S-cone response due to its
photopigment}
\item{s_mc}{S-cone-opic relating to the human M-cone response due to its
photopigment}
\item{s_lc}{S-cone-opic relating to the human L-cone response due to its
photopigment}
\item{s_rh}{rhodopic relating to the human rod response due to its
photopigment (rhodopsin)}
\item{s_mel}{melanopic relating to the human rod response due to its
photopigment (melanopsin)}
}
}
\source{
\insertRef{CIE2018}{lighting}
}
\usage{
cie2018_S_026
}
\description{
A dataset containing the action spectra representing the relative spectral
sensitivity of the five alpha-opic human photoreceptors to optical radiation
incident at the cornea, normalised to have a maximum value of 1. These action
spectra are based on pre-receptoral filtering for a reference observer of age
32 years (CIE 1931 standard colorimetric observer) and are use to compute the
metrology of optical readiation for ipRGC-influenced responses to light.
}
\keyword{datasets}
