% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_colour_rendering.R
\name{compute_colour_rendering}
\alias{compute_colour_rendering}
\title{Title}
\usage{
compute_colour_rendering(spectrum, wavelength, cmf2, Km = 683.002, R = 1, S)
}
\arguments{
\item{S}{}
}
\description{
Title
}
