% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lighting.R
\name{compute_lighting}
\alias{compute_lighting}
\title{Computes lighting values from given spectra dataframe}
\usage{
compute_lighting(spectra, str_wavelength = NULL)
}
\arguments{
\item{spectra}{dataframe of spectra with one wavelength column in nm}

\item{str_wavelength}{name of wavelength column. Default is NULL: The first
column is the wavelength column}
}
\value{
dataframe of lighting values used in lighting
}
\description{
Computes lighting values from given spectra dataframe
}
\examples{
wavelength <- seq(380, 780, 1)
P2700 <- planck_law(2700, wavelength)
P6500 <- planck_law(6500, wavelength)

spectra <- data.frame(wavelength, P2700, P6500)
spectra$EES <- 1

compute_lighting(spectra)
compute_lighting(spectra, "wavelength")

}
