% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cvrl.org}
\alias{cvrl.org}
\title{CIE luminosity and colour matching functions}
\format{
A wide format data frame with 4 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(360, 830, 1)}}
\item{V_pho}{CIE-1924 photopic luminosity function, normalised to a maximum
value of 1.}
\item{V_sco}{CIE-1951 scotopic luminosity function, normalised to a maximum
value of 1.}
\item{x_cmf_2}{CIE-1931 2-degree colour matching function x}
\item{y_cmf_2}{CIE-1931 2-degree colour matching function y}
\item{z_cmf_2}{CIE-1931 2-degree colour matching function z}
\item{x_cmf_10}{CIE-1964 10-degree colour matching function x}
\item{y_cmf_10}{CIE-1964 10-degree colour matching function y}
\item{z_cmf_10}{CIE-1964 10-degree colour matching function z}
}
}
\source{
\url{www.cvrl.org}
}
\usage{
cvrl.org
}
\description{
A dataset containing the CIE-1924 photopic luminosity function, CIE-1951
scotopic luminosity function, CIE-1931 2-degree XYZ colour matching functions
and CIE 1964 10-degree XYZ colour matching functions.
}
\keyword{datasets}
