% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{extract_lum_intensity_ldt}
\alias{extract_lum_intensity_ldt}
\title{Extract luminous intensity data from a LDT-file into a tibble}
\usage{
extract_lum_intensity_ldt(C, i, gamma, Ng, lines_data)
}
\arguments{
\item{C}{A single C-plane value.}

\item{i}{Counter for the unique C-plane values. Used for extracting the right
lines.}

\item{gamma}{A vector of gamma angles. Length corresponds to the number of
luminous intensity per gamma angle.}

\item{Ng}{Number of luminous intensities in each C-plane.}

\item{lines_data}{Lines with luminous intensity data from the LDT file.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with luminous intensities, C-planes and gamma angles.
}
\description{
\code{extract_lum_intensity_ldt()} returns a tibble of luminous
intensity data extracted from a LDT-file. The luminous intensities depends on
C-plane and gamma angles.
}
\details{
This function is used internally in the packages for iteration purposes.
}
