% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_spectrum.R
\name{integrate_spectrum}
\alias{integrate_spectrum}
\title{Integrates a dataframe of spectra}
\usage{
integrate_spectrum(spectrum, wavelength, sensitivity = 1, constant = 1)
}
\arguments{
\item{spectra}{}
}
\value{

}
\description{
Typical integration to compute lighting values
\deqn{E = K \cdot \int E_\text{e}(\lambda) \cdot s(\lambda) d\lambda}.
}
\examples{

}
