% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_spectra.R
\name{interpolate_spectra}
\alias{interpolate_spectra}
\title{Interpolates a dataframe of spectra with a wavelength column}
\usage{
interpolate_spectra(
  spectra,
  wl_out = seq(380, 780, 5),
  str_wavelength = NULL,
  method = "linear"
)
}
\arguments{
\item{spectra}{is a dataframe of spectra in wide table format including
a wavelength column in nm.}

\item{wl_out}{defines the output wavelength range and steps in nm for
interpolation.
Default: \code{wl_out = seq(380,780,5)}}

\item{str_wavelength}{Define name of wavelength column in \code{spectra}
dataframe. Default: \code{str_wavelength = NULL}. If entry is \code{NULL},
then the name of the first column from dataframe \code{spectra} will
chosen as \code{str_wavelength}.}

\item{method}{specifies the interpolation method of the functions
\link[stats]{approx} and \link[stats]{spline}. Default: \code{"linear"}.
Other choices for \link[stats]{spline} are: \code{"fmm"},
\code{"periodic"}, \code{"natural"}, \code{"monoH.FC"} and \code{"hyman"}.}
}
\value{
returns a dataframe of interpolated spectra in wide table format
  specified by \code{wl_out}.
}
\description{
Interpolates a dataframe of spectra with a wavelength column
}
\examples{
# Create spectrum planckian radiator using black body temperature in K
wavelength <- seq(380, 780, 5)
planck2700 <- colorscience::emittanceblackbodyPlanck(seq(380, 780, 5), 2700)
planck5000 <- colorscience::emittanceblackbodyPlanck(seq(380, 780, 5), 5000)
spectra <- data.frame(wavelength, planck2700, planck5000)
interpolate_spectra(spectra, seq(380, 780, 1), method = "linear")

}
\author{
William Truong
}
