% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_sprague.R
\name{interpolate_sprague}
\alias{interpolate_sprague}
\title{Interpolating using the Sprague method}
\usage{
interpolate_sprague(y, f)
}
\arguments{
\item{y}{numeric spectral data y giving to be interpolated}

\item{f}{interpolation factor. Doubled sampling rate (f = 2).}
}
\value{
returns interpolated data points using the sprague method.
}
\description{
Perform sprague interpolation of given spectrum data points and returning a
list of points obtained by interpolation. The original Matlab code by
\insertCite{Westland2012}{lighting} was adapted to R. Sampling rate must be
at least doubled, i. e. f \eqn{\ge}2. And f must be nominal (2, 3, ...).
For more insight in the interpolation method see
\insertCite{Westland2015}{lighting}.
}
\examples{
# Creating y-function as an alternative vor spectral data
x <- seq(-100, 100, 2)
y <- x^3
x_out <- seq(-100, 100, 0.5)
f <- 4 # -> steps: 2 / 0.5

interpolate_sprague(y, f)

}
\references{
\insertAllCited{}
}
