% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{ld_add_light_distribution}
\alias{ld_add_light_distribution}
\title{Adds plot polar chart into \code{ld_list}}
\usage{
ld_add_light_distribution(
  ld_list,
  line_color = "#BCCF03",
  line_size = 1.5,
  title = "",
  x_lab = expression("Drehwinkel" ~ gamma ~ "in °"),
  y_lab = expression(paste("Normierte Lichtstärke I in cd/1000 lm"))
)
}
\arguments{
\item{ld_list}{A specific light distribution list}

\item{line_color}{A text in hex colour format}

\item{line_size}{A numeric, which controls the line size of \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}}

\item{title}{The text for the plot title}

\item{x_lab}{The text for the x-axis label}

\item{y_lab}{The text for the y-axis label}
}
\value{
A \code{ld_list}, like \link{ld_data}
}
\description{
\code{ld_add_light_distribution()} adds the ggplot2 object from
\code{\link[=plot_light_distribution]{plot_light_distribution()}} into the light distribution list \code{ld_list}
(for list items description see \link{ld_data}).
}
\examples{
ld_add_light_distribution(ld_data)

}
