% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{ld_update}
\alias{ld_update}
\title{Update light distribution list}
\usage{
ld_update(
  ld_list,
  filepath = NA,
  file_name = NA,
  company = NA,
  report_no = NA,
  luminaire_name = NA,
  luminaire_no = NA,
  date_user = NA,
  length = NA,
  width = NA,
  height = NA,
  length_lum = NA,
  width_lum = NA,
  height_lum_C0 = NA,
  height_lum_C90 = NA,
  height_lum_C180 = NA,
  height_lum_C270 = NA,
  lamp_no = NA,
  lamp_type = NA,
  lum_flux = NA,
  cct = NA,
  cri = NA,
  power = NA
)
}
\arguments{
\item{ld_list}{A specific light distribution list}

\item{filepath}{Filepath of imported original light distribution file}

\item{file_name}{Name of LDT-file without extension}

\item{company}{Company identification/databank/version/format identification}

\item{report_no}{Measurement report number}

\item{luminaire_name}{Luminaire name}

\item{luminaire_no}{Luminaire number}

\item{date_user}{Date/user}

\item{length}{Length/diameter of luminaire (mm)}

\item{width}{b - Width of luminaire (mm) (b = 0 for circular luminaire)}

\item{height}{Height of luminaire (mm)}

\item{length_lum}{Length/diameter of luminous area (mm)}

\item{width_lum}{b1 - Width of luminous area (mm) (b1 = 0 for circular luminous area of luminaire)}

\item{height_lum_C0}{Height of luminous area C0-plane (mm)}

\item{height_lum_C90}{Height of luminous area C90-plane (mm)}

\item{height_lum_C180}{Height of luminous area C180-plane (mm)}

\item{height_lum_C270}{Height of luminous area C270-plane (mm)}

\item{lamp_no}{Number of lamps. For absolute photometry, number is
negative.}

\item{lamp_type}{Type of lamps}

\item{lum_flux}{Total luminous flux of lamps (lm). For absolute
photometry, this field is Total Luminous Flux of Luminaire.}

\item{cct}{Color appearance / color temperature of lamps}

\item{cri}{Color rendering group / color rendering index}

\item{power}{Wattage including ballast (W)}
}
\value{
\code{ld_update()} returns a light distribution list (ld_list)
}
\description{
\code{ld_update()} changes configurable parameters in the light
distribution data. The principal light distribution is not altered in the process.
}
\examples{
ld_update(ld_data, file = "test")
}
