% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{ld_write_ldt}
\alias{ld_write_ldt}
\title{Write light distribution data (\code{ld_list}) to a LDT file}
\usage{
ld_write_ldt(ld_list, file, user = "")
}
\arguments{
\item{ld_list}{A specific light distribution list}

\item{file}{The path to file without extension.}

\item{user}{The text naming the user, who manipulated the LDT-file}
}
\value{
\code{ld_write_ldt()} returns a NULL invisibly.
}
\description{
\code{ld_write_ldt()} exports light distribution data (\code{ld_list}) as
an EULUMDAT file (*.ldt).
}
\examples{
# ld_write_ldt() will write the file to the current working directory, if
# it is unspecified without directory path.
\dontrun{
ld_write_ldt(ld_data, file = "test")
}
}
