% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting-package.R
\docType{package}
\name{lighting-package}
\alias{lighting}
\alias{lighting-package}
\title{lighting: Lighting Tools}
\description{
Various tools used in the lighting industry. Reading, writing and plotting light distribution files *.ldt and *.ies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Wei-Lim/lighting}
  \item Report bugs at \url{https://github.com/Wei-Lim/lighting/issues}
}

}
\author{
\strong{Maintainer}: William Truong \email{willi.truong@gmail.com} (\href{https://orcid.org/0000-0002-9857-4437}{ORCID})

}
\keyword{internal}
