% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lucas2014}
\alias{lucas2014}
\title{Spectral sensitivity function of human photopigments}
\format{
A wide format data frame with 4 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(380, 780, 5)}}
\item{N_sc}{Cyanolabe response function (S-cone photopsin), normalised to
give an integrated area of 1}
\item{N_mc}{Chlorolabe response function (M-cone photopsin), normalised to
give an integrated area of 1}
\item{N_lc}{Erythrolabe response function (L-cone photopsin), normalised to
give an integrated area of 1}
\item{N_rh}{Rod opsin response function (Rod opsin), normalised to
give an integrated area of 1}
\item{N_z}{Melanopsin response function (Melanopsin), normalised to
give an integrated area of 1}
\item{V_pho}{CIE-1924 photopic luminosity function, normalised to a maximum
value of 1.}
\item{T_lens}{Eye lens transmission of a 32 year old standard observer.}
}
}
\source{
\insertRef{Lucas2014}{lighting}
}
\usage{
lucas2014
}
\description{
A dataset containing the spectral sensitivity function of human
photopigments, photopic luminosity function and eye lens transmission
function provided by \insertCite{Lucas2014}{lighting} to calculate the
alpha-opic illuminances.
}
\keyword{datasets}
