% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planck_law.R
\name{planck_law}
\alias{planck_law}
\title{Computes spectrum of a Planckian radiator}
\usage{
planck_law(temperature, wl)
}
\arguments{
\item{temperature}{of black body radiator in kelvin}

\item{wl}{wavelength array in nm of exported spectrum}
}
\value{
spectrum of Planckian radiator
}
\description{
Computes spectrum of a Planckian radiator
}
\examples{
# create wavelength array
wl <- seq(380, 780, 1)

planck_law(2700, wl)
planck_law(6500, wl)
}
