% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rea2010}
\alias{rea2010}
\title{Spectral sensitivity functions to calculate Circadian Stimulus (2010)}
\format{
A wide format data frame with 4 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(380, 780, 10)}}
\item{S_sc}{S-cone spectral efficiency function}
\item{V_sco}{Rod spectral efficiency function}
\item{M}{Melanopsin-containing retinal ganglion cell spectral
efficiency function peaking at 480 nm}
\item{V_pho_10}{Large-filed L+M cone spectral efficiency function}
}
}
\source{
\insertRef{Rea2010}{lighting}
}
\usage{
rea2010
}
\description{
A dataset containing the spectral sensitivity function provided by
\insertCite{Rea2010}{lighting} to calculate the Circadian Stimulus
(deprecated version 2010).
}
\keyword{datasets}
