% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rea2018}
\alias{rea2018}
\title{Spectral sensitivity functions to calculate Circadian Stimulus (2018)}
\format{
A wide format data frame with 4 variables:
\describe{
\item{nm}{wavelength, in nm. Range and intervall: \code{seq(380, 780, 2)}}
\item{V_pho}{photopic luminous efficiency function}
\item{V_sco}{scotopic luminous efficiency function}
\item{V_pho_mac}{photopic luminous efficiency function weighted with
macular pigment transmittance}
\item{S_sc_mac}{S-cone fundamental weighted with macular pigment
transmittance}
\item{Mc_lens}{Melanopsin spectral effiency function, corrected for
crystalline lens transmittance}
}
}
\source{
\insertRef{Rea2018}{lighting} \cr
\url{http://www.lrc.rpi.edu/resources/CSCalculator_2017_10_03_Mac.xlsm}
}
\usage{
rea2018
}
\description{
A dataset containing the spectral sensitivity function provided by
\insertCite{Rea2018}{lighting} to calculate the Circadian Stimulus
(2018).
}
\keyword{datasets}
