% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{read_ldt}
\alias{read_ldt}
\title{Read a LDT-file into a specific light distribution list format}
\usage{
read_ldt(file)
}
\arguments{
\item{file}{Path to LDT-file.}
}
\value{
A specific light distribution list (\code{ld_list}), that can be
used with all functions that have a prefix \code{ld_} (for list items description
see \link{ld_data}).
}
\description{
\code{read_ldt()} is used to extract light distribution data of a
LDT-file into a specific list format \code{ld_list} (for item description see \link{ld_data}).
}
\examples{
library(tidyverse)
library(furrr)

file <- system.file('extdata', "ldt_min_example.ldt", package = "lighting")

read_ldt(file)

}
