% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_values.R
\name{sensitivity_functions}
\alias{sensitivity_functions}
\title{Returns list of interpolated spectral sensitivity functions}
\usage{
sensitivity_functions(wl_out = seq(380, 780, 1))
}
\arguments{
\item{wl_out}{defines the wavelength interval. Wavelength range
(380-780)nm.}
}
\value{
list of sensitivity functions for computing lighting values
}
\description{
In order to compute lighting values by numerical intergration the spectral
power distribution and spectral sensitivity function must be interpolated to
a common wavelength range and interval. This function provides a collection
sensitivity function for calcuation of different lighting values. As
recommended by \insertCite{CIE2018b}{lighting} the sprague method is
performed to interpolate smooth spectra. The exception is the
interpolation of daylight components and CIE test colour samples, which
should be explicitly interpolated linearly according to
\insertCite{CIE2018b,CIE1995}{lighting}.
}
\details{
Firstly, the sensitivity functions are interpolated to specified wavelength
interval and their original wavelength range using the recommended methods by
\insertCite{CIE2018b,CIE1995}{lighting}. Secondly the functions are linear
interpolated to the specified wavelength range and interval; extrapolated
values are returned with 0. Wavelength interval cannot be smaller than 1.
}
\examples{
# Returning in wavelength seq(380, 780, 1)
sensitivity_functions()
# Returning original datasets as a list
sensitivity_functions(NULL)
# Returning in wavelength seq(200, 830, 1)
sensitivity_functions(seq(200, 830, 1))

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\\insertAllCited\{\}
}\if{html}{\out{</div>}}
}
