% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{extract_lum_intensity_ldt}
\alias{extract_lum_intensity_ldt}
\title{Extracts Luminous Intensity LDT Data}
\usage{
extract_lum_intensity_ldt(C, i, gamma, Ng, lines_data)
}
\arguments{
\item{C}{A single C-plane value.}

\item{i}{Counter for the unique C-plane values. Used for extracting the right
lines.}

\item{gamma}{A vector of gamma angles. Length corresponds to the number of
luminous intensity per gamma angle.}

\item{Ng}{Number of luminous intensities in each C-plane.}

\item{lines_data}{Lines with luminous intensity data from the LDT file.}
}
\description{
Extracts luminous intensity from LDT-file. It depends on the number of
luminous intensities per C-plane and per gamma angle. Internal private
function for iteration purposes.
}
