% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_values.R
\name{interpolate_sprague}
\alias{interpolate_sprague}
\title{Interpolating spectrum using the Sprague method}
\usage{
interpolate_sprague(y, f)
}
\arguments{
\item{y}{Numerical spectral data vector to be interpolated}

\item{f}{Interpolation factor. Doubling the sampling rate with f = 2.}
}
\value{
A vector of interpolated data points using the Sprague method.
}
\description{
\code{interpolate_sprague()} performs a special interpolation for a
spectrum with equidistant data points. interpolation of defined
spectrum data points.
}
\details{
This function is originally based on a Matlab code \insertCite{Westland2012}{lighting}
and was adapted to R. Sampling rate must be at least doubled, i. e. f \eqn{\ge}2.
And f must be nominal (2, 3, ...). #' For more insight of the sprague
interpolation method see \insertCite{Westland2015}{lighting}.
}
\examples{
spectra_tbl <- tibble::tibble(wavelength = seq(380, 780, 5)) \%>\%
	dplyr::mutate(p2700K = planck_law(wavelength, 2700)) \%>\%
	dplyr::mutate(p5000K = planck_law(wavelength, 5000))

# interpolation to wavelength range seq(380, 780, 5)
interpolate_sprague(spectra_tbl$p2700K, 5)

}
\references{
\if{html}{\out{<div class="sourceCode">}}\preformatted{\\insertAllCited\{\}
}\if{html}{\out{</div>}}
}
