% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{ld_write_ies_lm63_2002}
\alias{ld_write_ies_lm63_2002}
\title{Write light distribution data (\code{ld_list}) to an IES file}
\usage{
ld_write_ies_lm63_2002(ld_list, dir_path = "")
}
\arguments{
\item{ld_list}{A specific light distribution list}

\item{dir_path}{The path to file directory. The filename will provided by
\code{ld_list}, see \link{ld_data} for item descriptions.}
}
\value{
\code{ld_write_ies_lm63_2002()} returns a NULL invisibly.
}
\description{
\code{ld_write_ies_lm63_2002()} exports light distribution data (\code{ld_list})
after ANSI/IESNA LM-63-2002 standard in IES file format (*.ies).
}
\examples{
# ld_write_ies_lm63_2002() will write the file to the current working
# directory, if dir_path is unspecified.
\dontrun{
ld_write_ies_lm63_2002(ld_data)
}
}
