% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lighting_distribution.R
\name{ld_write_svg}
\alias{ld_write_svg}
\title{Write a \code{ld_list} polar chart to a SVG file}
\usage{
ld_write_svg(ld_list, dir_path)
}
\arguments{
\item{ld_list}{A specific light distribution list}

\item{dir_path}{The path to file directory. The filename will provided by
\code{ld_list}, see \link{ld_data} for item descriptions.}
}
\value{
\code{ld_write_svg()} returns the ggplot2 object invisibly.
}
\description{
\code{ld_write_svg()} exports the ggplot2 object of the specific
light distritbution list \code{ld_list} as a SVG graphic file.
}
\examples{
# If only a file name is specified, ld_write_svg() will write
# the file to the current working directory.
\dontrun{
ld_write_svg(ld_data, dir_path = "")
}
}
